Add query validation configuration for quickstart (#3040)
Currently all attempts to use the Query/{logicName}/validate endpoint to
validate queries in the Quickstart will result in a 'Query validation
not implemented' error due to the lack of configured rules for the
baseline ShardQueryLogic bean.

Add default query validation rule definitions to the quickstart
QueryLogicFactory.xml to fix this.