Bugfix/root pointer (#3149)
This fixes a bug where more than the root / TLD document was being
returned to the user with a luceneUUIDLookup. The problem was that the
UID of the document in question was actually smaller than normal which
broke some optimizations. Now that the event data filters can maintain
context of the document, I rewrote this to more efficiently determine
whether a key is for a root document or not.