Initial commit of base annotation data model and README (#3169)
Provides a generic method to annotate portions of data stored in
Datawave. Will provide a data model, accumulo schema and table
definition and a web API for create, read, update and delete operations
for annotations.

This PR contains the base protocol buffers files for this feature and
the generated java source code and json schema. Future PR's will build
upon this foundation and address the scope of items covered in the
README