Add Prediction.confidence property (#3110)
Add `confidence` property to the BaseQueryMetric.Prediction class in
anticpation of needing to include the accuracy of the prediction. A
separate PR will be created to update the usage of Prediction in the
QueryExecutorBean once a new release version of query-metrics has been
created with the updates here.

Part of work for #3099